Adding project code
These are the C# scripts for making the simulation run. Note that this is code only, and much of it requires objects to be created for the simulation to work.
If you are interested in testing the simulation itself, that is uploaded as a zip file that you may download and test. Feel free to modify, add to, or play around with the simulation as you please, as many of the settings are modifiable to simulate different situations.